requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
celerybeat-mongo==0.2.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.0.8
django-celery-beat==2.4.0
django-celery-results==2.4.0
django-cors-headers==3.13.0
django-cron==0.6.0
django-crontab==0.7.1
django-request-logging==0.7.5
django-timezone-field==5.0
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.2.1
kombu==5.2.4
mongoengine==0.24.2
prompt-toolkit==3.0.36
pymongo==3.12.3
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2022.6
redis==4.4.0
six==1.16.0
sqlparse==0.2.4
tzdata==2022.7
vine==5.0.0
wcwidth==0.2.5
